Cast off_t to long long before printing. Use %ll instead of %q incase long long and quad_t are not the same.
git-svn-id: svn+ssh://svn.freebsd.org/base/head@87282 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f

+++ usr.bin/cmp/special.c
@@ -80,7 +80,8 @@ c_special(fd1, file1, skip1, fd2, file2, skip2)
 		if (ch1 != ch2) {
 			if (lflag) {
 				dfound = 1;
-				(void)printf("%6qd %3o %3o\n", byte, ch1, ch2);
+				(void)printf("%6lld %3o %3o\n",
+				    (long long)byte, ch1, ch2);
 			} else {
 				diffmsg(file1, file2, byte, line);
 				/* NOTREACHED */
